requirements.txt
Django~=4.2.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
langchain~=0.0.285